backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
billiard==4.2.3
black==23.11.0
boto3==1.29.7
botocore==1.32.7
cachetools==6.2.2
cbor2==5.7.1
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.0
django-filter==23.3
django-ses==3.5.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==20.1.0
flake8==6.1.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.3.1
google-genai==1.52.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
isort==5.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==1.3.7
packaging==25.0
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.1.0
PyJWT==2.8.0
pyOpenSSL==25.3.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.8.0
sentry-sdk==1.38.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
zope.interface==8.1.1
